corrected POST API endpoint
