+++ galasa-parent/dev.galasa.framework.api.common/src/testFixtures/java/dev/galasa/framework/api/common/mocks/MockIConfigurationPropertyStoreService.java
@@ -23,19 +23,19 @@ public class MockIConfigurationPropertyStoreService implements IConfigurationPro
 
     public MockIConfigurationPropertyStoreService(@NotNull String namespace) {
         this.namespaceInput = namespace;
-        if (this.namespaceInput != "multi"){
-            this.properties.put("property1", "value1");
-            this.properties.put("property2", "value2");
-            this.properties.put("property3", "value3");
-            this.properties.put("property4", "value4");
-            this.properties.put("property5", "value5");
-        }else{
+        if (this.namespaceInput == "multi"){
             this.properties.put("test.property", "value1");
             this.properties.put(".charity1", "value2");
             this.properties.put(".lecture101", "value101");
             this.properties.put(".hospitality", "value3");
             this.properties.put("test.aunty5", "value4");
             this.properties.put("test.empty", "value5");
+        }else{
+            this.properties.put("property1", "value1");
+            this.properties.put("property2", "value2");
+            this.properties.put("property3", "value3");
+            this.properties.put("property4", "value4");
+            this.properties.put("property5", "value5");
         }
 	}
