+++ galasa-parent/dev.galasa.framework.api.cps/src/test/java/dev/galasa/framework/api/cps/internal/routes/TestPropertyRoute.java
@@ -10,6 +10,7 @@ import dev.galasa.framework.api.cps.internal.mocks.MockCpsServlet;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.framework;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -20,16 +21,15 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.junit.Test;
 
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParser;
+
+
 public class TestPropertyRoute extends CpsServletTest{
-
-    
-    /*
-     * TESTS  -- GET requests
-     */
     @Test
-    public void TestPropertyRouteGETNoFrameworkReturnsError() throws Exception{
+    public void TestPropertyQueryNoFrameworkReturnsError() throws Exception{
 		// Given...
-		setServlet("/cps/namespace1/properties/property1",null ,new HashMap<String,String[]>());
+		setServlet("/cps/namespace1/properties",null ,new HashMap<String,String[]>());
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -40,48 +40,23 @@ public class TestPropertyRoute extends CpsServletTest{
 		servlet.doGet(req,resp);
 
 		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
+		// We expect an error back, because the API server couldn't find any Etcd store to query
+		assertThat(resp.getStatus()==500);
 		assertThat(resp.getContentType()).isEqualTo("application/json");
 		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
 
 		checkErrorStructure(
 			outStream.toString(),
-			5016,
-			"GAL5016E: Error occured when trying to access namespace 'namespace1'. The namespace provided is invalid."
+			5000,
+			"GAL5000E: ",
+			"Error occured when trying to access the endpoint"
 		);
     }
 
     @Test
-    public void TestPropertyRouteGETBadNamespaceReturnsError() throws Exception{
-		// Given...
-		setServlet("/cps/error/properties/property1",null ,new HashMap<String,String[]>());
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-				
-		// When...
-		servlet.init();
-		servlet.doGet(req,resp);
-
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5016,
-			"E: Error occured when trying to access namespace 'error'. The namespace provided is invalid."
-		);
-    }
-
-    @Test
-    public void TestPropertyRouteGETWithExistingNamespaceReturnsOk() throws Exception {
+    public void TestPropertyQueryWithExistingNamespaceReturnsOk() throws Exception {
         // Given...
-        setServlet("/cps/framework/properties/property1", "framework", new HashMap<String,String[]>());
+        setServlet("/cps/framework/properties", "framework", new HashMap<String,String[]>());
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -92,6 +67,10 @@ public class TestPropertyRoute extends CpsServletTest{
         servlet.doGet(req, resp);
         Map<String, String> properties = new HashMap<String,String>();
         properties.put("property1", "value1");
+        properties.put("property2", "value2");
+        properties.put("property3", "value3");
+        properties.put("property4", "value4");
+        properties.put("property5", "value5");
 
         // Then...
         // We expect data back
@@ -100,12 +79,13 @@ public class TestPropertyRoute extends CpsServletTest{
 		assertThat(resp.getContentType()).isEqualTo("application/json");
 		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
 		checkJsonArrayStructure(output,properties);
+        
     }
 
-    @Test
-    public void TestPropertyRouteGETWithExistingNamespaceDifferentPropertyReturnsOk() throws Exception {
+	@Test
+    public void TestPropertyQueryWithProtectedNamespaceReturnsOk() throws Exception {
         // Given...
-        setServlet("/cps/framework/properties/property3", "framework", new HashMap<String,String[]>());
+        setServlet("/cps/secure/properties", "secure", new HashMap<String,String[]>());
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -115,7 +95,11 @@ public class TestPropertyRoute extends CpsServletTest{
         servlet.init();
         servlet.doGet(req, resp);
         Map<String, String> properties = new HashMap<String,String>();
-        properties.put("property3", "value3");
+        properties.put("property1", "********");
+        properties.put("property2", "********");
+        properties.put("property3", "********");
+        properties.put("property4", "********");
+        properties.put("property5", "********");
 
         // Then...
         // We expect data back
@@ -124,12 +108,203 @@ public class TestPropertyRoute extends CpsServletTest{
 		assertThat(resp.getContentType()).isEqualTo("application/json");
 		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
 		checkJsonArrayStructure(output,properties);
+        
     }
 
     @Test
-    public void TestPropertyRouteGETWithExistingNamespaceBadPropertyNameReturnsEmpty() throws Exception {
+    public void TestPropertyQueryHiddenNamespaceReturnsError() throws Exception {
         // Given...
-        setServlet("/cps/framework/properties/inproperty", "framework", new HashMap<String,String[]>());
+		setServlet("/cps/dss/properties", "dss" ,new HashMap<String,String[]>());
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+		ServletOutputStream outStream = resp.getOutputStream();	
+				
+		// When...
+		servlet.init();
+		servlet.doGet(req,resp);
+
+		// Then...
+		// We expect an error back, because the API server could find the namespace, but it was hidden
+		assertThat(resp.getStatus()==500);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+
+		checkErrorStructure(
+			outStream.toString(),
+			5016,
+			"GAL5016E: ",
+			"Error occured when trying to access namespace 'dss'. The namespace provided is invalid"
+		);
+    }
+
+	@Test
+    public void TestPropertyQueryInvalidNamespaceReturnsError() throws Exception {
+        // Given...
+		setServlet("/cps/j!ndex/properties", "framework" ,new HashMap<String,String[]>());
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+		ServletOutputStream outStream = resp.getOutputStream();	
+				
+		// When...
+		servlet.init();
+		servlet.doGet(req,resp);
+
+		// Then...
+		// We expect an error back, because the API server could find the namespace
+		assertThat(resp.getStatus()==500);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+
+		checkErrorStructure(
+			outStream.toString(),
+			5404,
+			"GAL5404E:",
+			" Error occured when trying to identify the endpoint '/cps/j!ndex/properties'. ",
+			"Please check your endpoint URL or report the problem to your Galasa Ecosystem owner."
+		);
+	}
+
+	@Test
+	public void TestGetPropertiesWithSuffixNoMatchReturnsEmpty() {
+		//Given...
+		String suffix  = "rty1";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value2");
+		properties.put("property3", "value3");
+		properties.put("property4", "value4");
+		properties.put("property5", "value5");
+		properties.put("property6", "value6");
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesBySuffix(properties, suffix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+	@Test
+	public void TestGetPropertiesWithSuffixReturnsOneRecord() {
+		//Given...
+		String suffix  = "1";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+		expectedProperties.put("property1", "value1");
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value2");
+		properties.put("property3", "value3");
+		properties.put("property4", "value4");
+		properties.put("property5", "value5");
+		properties.put("property6", "value6");
+		properties.putAll(expectedProperties);
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesBySuffix(properties, suffix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+	@Test
+	public void TestGetPropertiesWithSuffixReturnsFiveRecord() {
+		//Given...
+		String suffix  = "ty";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+		expectedProperties.put("property", "value1");
+		expectedProperties.put("charity", "value2");
+		expectedProperties.put("hospitality", "value3");
+		expectedProperties.put("aunty", "value4");
+		expectedProperties.put("empty", "value5");
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value6");
+		properties.put("property3", "value7");
+		properties.put("property4", "value8");
+		properties.put("property5", "value9");
+		properties.put("property6", "value10");
+		properties.putAll(expectedProperties);
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesBySuffix(properties, suffix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+	@Test
+	public void TestGetPropertiesWithPrefixNoMatchReturnsEmpty() {
+		//Given...
+		String prefix  = "crate";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value2");
+		properties.put("property3", "value3");
+		properties.put("property4", "value4");
+		properties.put("property5", "value5");
+		properties.put("property6", "value6");
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesByPrefix(properties, prefix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+	@Test
+	public void TestGetPropertiesWithPrefixReturnsOneRecord() {
+		//Given...
+		String prefix  = "pre";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+		expectedProperties.put("preperty1", "value1");
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value2");
+		properties.put("property3", "value3");
+		properties.put("property4", "value4");
+		properties.put("property5", "value5");
+		properties.put("property6", "value6");
+		properties.putAll(expectedProperties);
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesByPrefix(properties, prefix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+	@Test
+	public void TestGetPropertiesWithPrefixReturnsFiveRecord() {
+		//Given...
+		String prefix  = ".";
+		Map<String, String> expectedProperties = new HashMap<String,String>();
+		expectedProperties.put(".property", "value1");
+		expectedProperties.put(".charity", "value2");
+		expectedProperties.put(".hospitality", "value3");
+		expectedProperties.put(".aunty", "value4");
+		expectedProperties.put(".empty", "value5");
+		Map<String, String> properties = new HashMap<String,String>();
+		properties.put("property2", "value6");
+		properties.put("property3", "value7");
+		properties.put("property4", "value8");
+		properties.put("property5", "value9");
+		properties.put("property6", "value10");
+		properties.putAll(expectedProperties);
+
+		//When...
+		Map<String, String> results = new PropertyRoute(null,null).filterPropertiesByPrefix(properties, prefix);
+		
+		//Then...
+		assertThat(results).isEqualTo(expectedProperties);
+	}
+
+    @Test
+    public void TestPropertyQueryWithNamespaceAndURLQuerySuffixReturnsOneRecord() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("suffix", new String[] {"1"});
+
+        setServlet("/cps/framework/properties?suffix=1", "framework", params);
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -140,17 +315,21 @@ public class TestPropertyRoute extends CpsServletTest{
         servlet.doGet(req, resp);
 
         // Then...
+        // We expect data back
         String output = outStream.toString();
         assertThat(resp.getStatus()).isEqualTo(200);
 		assertThat(resp.getContentType()).isEqualTo("application/json");
 		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-		assertThat(output).isEqualTo("[]");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \"property1\",\n    \"value\": \"value1\"\n  }\n]");
     }
 
-    @Test
-    public void TestPropertyRouteGETWithExistingNamespaceIncopmpletePropertyNameReturnsEmpty() throws Exception {
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithSuffixThatDoesNotExistReturnsEmpty() throws Exception {
         // Given...
-        setServlet("/cps/framework/properties/roperty", "framework", new HashMap<String,String[]>());
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("suffix", new String[] {"mickey"});
+
+        setServlet("/cps/framework/properties?suffix=mickey", "framework", params);
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -161,6 +340,7 @@ public class TestPropertyRoute extends CpsServletTest{
         servlet.doGet(req, resp);
 
         // Then...
+        // We expect data back
         String output = outStream.toString();
         assertThat(resp.getStatus()).isEqualTo(200);
 		assertThat(resp.getContentType()).isEqualTo("application/json");
@@ -168,14 +348,194 @@ public class TestPropertyRoute extends CpsServletTest{
 		assertThat(output).isEqualTo("[]");
     }
 
-    /*
-     * TESTS  -- PUT requests
-     */
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithSuffixReturnsMultipleRecords() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("suffix", new String[] {"ty"});
 
-    @Test
-    public void TestPropertyRoutePUTNoFrameworkReturnsError() throws Exception{
+        setServlet("/cps/multi/properties?suffix=mickey", "multi", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \".hospitality\",\n    \"value\": \"value3\"\n  },"+
+			"\n  {\n    \"name\": \"test.property\",\n    \"value\": \"value1\"\n  },\n  {\n    \"name\": \"test.empty\",\n    \"value\": \"value5\"\n  }\n]");
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryPrefixReturnsOneRecord() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {".char"});
+
+        setServlet("/cps/multi/properties?prefix=.char", "multi", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \".charity1\",\n    \"value\": \"value2\"\n  }\n]");
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixThatDoesNotExistReturnsEmpty() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {"goof"});
+
+        setServlet("/cps/framework/properties?prefix=goof", "framework", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[]");
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixReturnsMultipleRecords() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {"test"});
+
+        setServlet("/cps/multi/properties?prefix=test", "multi", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \"test.property\",\n    \"value\": \"value1\"\n  },"+
+			"\n  {\n    \"name\": \"test.empty\",\n    \"value\": \"value5\"\n  },\n  {\n    \"name\": \"test.aunty5\",\n    \"value\": \"value4\"\n  }\n]" );
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryPrefixAndSuffixReturnsOneRecord() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {"prop"});
+		params.put("suffix", new String[] {"5"});
+
+        setServlet("/cps/framework/properties?prefix=prop&suffix=5", "framework", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \"property5\",\n    \"value\": \"value5\"\n  }\n]");
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixAndSuffixThatDoesNotExistReturnsEmpty() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {"hello"});
+		params.put("suffix", new String[] {"world"});
+
+        setServlet("/cps/framework/properties?prefix=hellosuffix=world", "framework", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[]");
+    }
+
+	@Test
+    public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixAndSuffixReturnsMultipleRecords() throws Exception {
+        // Given...
+		Map <String,String[]> params = new HashMap<String,String[]>();
+		params.put("prefix", new String[] {"."});
+		params.put("suffix", new String[] {"1"});
+
+        setServlet("/cps/multi/properties?prefix=.&suffix=1", "multi", params);
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doGet(req, resp);
+
+        // Then...
+        // We expect data back
+        String output = outStream.toString();
+        assertThat(resp.getStatus()).isEqualTo(200);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+		assertThat(output).isEqualTo("[\n  {\n    \"name\": \".charity1\",\n    \"value\": \"value2\"\n  },"+
+			"\n  {\n    \"name\": \".lecture101\",\n    \"value\": \"value101\"\n  }\n]");
+    }
+
+	/*
+	 * TEST - HANDLE PUT REQUEST - should error as this method is not supported by this API end-point
+	 */
+	@Test
+	public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixAndSuffixPUTRequestsReturnsError() throws Exception{
 		// Given...
-		setServlet("/cps/namespace1/properties/property1",null ,"value6", "PUT");
+		setServlet("/cps/multi/properties?prefix=.&suffix=1","multi",null, "PUT");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -184,6 +544,64 @@ public class TestPropertyRoute extends CpsServletTest{
 		// When...
 		servlet.init();
 		servlet.doPut(req,resp);
+
+		// Then...
+		// We expect an error back, because the API server has thrown a ConfigurationPropertyStoreException
+		assertThat(resp.getStatus()).isEqualTo(405);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+
+		checkErrorStructure(
+			outStream.toString(),
+			5405,
+			"E: Error occured when trying to access the endpoint '/cps/multi/properties?prefix=.&suffix=1'. The method 'PUT' is not allowed."
+		);
+    }
+
+	/*
+	 * TEST - HANDLE DELETE REQUEST - should error as this method is not supported by this API end-point
+	 */
+	@Test
+	public void TestPropertyQueryWithNamespaceAndURLQueryWithPrefixAndSuffixDELETERequestsReturnsError() throws Exception{
+		// Given...
+		setServlet("/cps/multi/properties?prefix=.&suffix=1","multi",null, "DELETE");
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+		ServletOutputStream outStream = resp.getOutputStream();	
+				
+		// When...
+		servlet.init();
+		servlet.doDelete(req,resp);
+
+		// Then...
+		// We expect an error back, because the API server has thrown a ConfigurationPropertyStoreException
+		assertThat(resp.getStatus()).isEqualTo(405);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+
+		checkErrorStructure(
+			outStream.toString(),
+			5405,
+			"E: Error occured when trying to access the endpoint '/cps/multi/properties?prefix=.&suffix=1'. The method 'DELETE' is not allowed."
+		);
+    }
+
+	/*
+	 * TEST - HANDLE POST REQUEST
+	 */
+	@Test
+    public void TestPropertyRoutePOSTNoFrameworkReturnsError() throws Exception{
+		// Given...
+		setServlet("/cps/namespace1/properties",null ,"{name: property ,value: value }", "POST");
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+		ServletOutputStream outStream = resp.getOutputStream();	
+				
+		// When...
+		servlet.init();
+		servlet.doPost(req,resp);
 
 		// Then...
 		// We expect an error back, because the API server couldn't find any Etcd store to Route
@@ -199,9 +617,9 @@ public class TestPropertyRoute extends CpsServletTest{
     }
 
     @Test
-    public void TestPropertyRoutePUTBadNamespaceReturnsError() throws Exception{
+    public void TestPropertyRoutePOSTBadNamespaceReturnsError() throws Exception{
 		// Given...
-		setServlet("/cps/error/properties/property1",null ,"value6", "PUT");
+		setServlet("/cps/error/properties",null ,"{name: property ,value: value }", "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -209,7 +627,7 @@ public class TestPropertyRoute extends CpsServletTest{
 				
 		// When...
 		servlet.init();
-		servlet.doPut(req,resp);
+		servlet.doPost(req,resp);
 
 		// Then...
 		// We expect an error back, because the API server couldn't find any Etcd store to Route
@@ -225,11 +643,14 @@ public class TestPropertyRoute extends CpsServletTest{
     }
 
     @Test
-    public void TestPropertyRouteWithExistingNamespacePUTNewPropertyReturnsSuccess() throws Exception {
+    public void TestPropertyRouteWithExistingNamespacePOSTNewPropertyReturnsSuccess() throws Exception {
         // Given...
+		String namespace = "framework";
         String propertyName = "property6";
         String value = "value6";
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "PUT");
+		JsonElement requestJson = JsonParser.parseString(String.format("{name: %s ,value: %s }",propertyName,value));
+		String requestBody = requestJson.toString();
+		setServlet("/cps/framework/properties", namespace, requestBody , "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -237,7 +658,7 @@ public class TestPropertyRoute extends CpsServletTest{
 
         // When...
         servlet.init();
-        servlet.doPut(req, resp);
+        servlet.doPost(req, resp);
 
         // Then...
         Integer status = resp.getStatus();
@@ -246,15 +667,18 @@ public class TestPropertyRoute extends CpsServletTest{
 		assertThat(resp.getContentType()).isEqualTo("application/json");
 		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
         assertThat(output).isEqualTo("Successfully created property property6 in framework");
-        assertThat(checkNewPropertyInNamespace(propertyName, value)).isTrue();       
+        assertThat(checkNewPropertyInNamespace(namespace, propertyName, value)).isTrue();       
     }
 
-    @Test
-    public void TestPropertyRouteWithExistingNamespacePUTExistingPropertyReturnsError() throws Exception {
+	@Test
+    public void TestPropertyRouteWithProtectedNamespacePOSTNewPropertyReturnsSuccess() throws Exception {
         // Given...
-        String propertyName = "property5";
+		String namespace = "secure";
+        String propertyName = "property6";
         String value = "value6";
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "PUT");
+		JsonElement requestJson = JsonParser.parseString(String.format("{name: %s ,value: %s }",propertyName,value));
+		String requestBody = requestJson.toString();
+		setServlet("/cps/secure/properties", namespace, requestBody , "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -262,7 +686,63 @@ public class TestPropertyRoute extends CpsServletTest{
 
         // When...
         servlet.init();
-        servlet.doPut(req, resp);
+        servlet.doPost(req, resp);
+
+        // Then...
+        Integer status = resp.getStatus();
+        String output = outStream.toString();
+        assertThat(status).isEqualTo(201);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+        assertThat(output).isEqualTo("Successfully created property property6 in secure");
+        assertThat(checkNewPropertyInNamespace(namespace, propertyName, value)).isTrue();       
+    }
+
+	@Test
+    public void TestPropertyRouteWithHiddenNamespacePOSTNewPropertyReturnsError() throws Exception {
+        // Given...
+        String propertyName = "property6";
+        String value = "value6";
+		JsonElement requestJson = JsonParser.parseString(String.format("{name: %s ,value: %s }",propertyName,value));
+		String requestBody = requestJson.toString();
+		setServlet("/cps/dss/properties", "framework", requestBody , "POST");
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doPost(req, resp);
+
+        // Then...
+        assertThat(resp.getStatus()).isEqualTo(404);
+		assertThat(resp.getContentType()).isEqualTo("application/json");
+		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
+
+		checkErrorStructure(
+			outStream.toString(),
+			5016,
+			"GAL5016E: Error occured when trying to access namespace 'dss'. The namespace provided is invalid."
+		);    
+    }
+
+    @Test
+    public void TestPropertyRouteWithExistingNamespacePOSTExistingPropertyReturnsError() throws Exception {
+        // Given...
+        String propertyName = "property5";
+        String value = "value6";
+		JsonElement requestJson = JsonParser.parseString(String.format("{name: %s ,value: %s }",propertyName,value));
+		String requestBody = requestJson.toString();
+        setServlet("/cps/framework/properties", "framework", requestBody, "POST");
+		MockCpsServlet servlet = getServlet();
+		HttpServletRequest req = getRequest();
+		HttpServletResponse resp = getResponse();
+        ServletOutputStream outStream = resp.getOutputStream();	
+
+        // When...
+        servlet.init();
+        servlet.doPost(req, resp);
 
         // Then...
         Integer status = resp.getStatus();
@@ -273,16 +753,18 @@ public class TestPropertyRoute extends CpsServletTest{
 			outStream.toString(),
 			5018,
 			"E: Error occured when trying to access property 'property5'.",
-            " The property name provided already exists in the 'framework' namesapce."
+            " The property name provided already exists in the 'framework' namespace."
 		);        
     }
 
     @Test
-    public void TestPropertyRouteWithErroneousNamespacePUTNewPropertyReturnsError() throws Exception {
+    public void TestPropertyRouteWithErroneousNamespacePOSTNewPropertyReturnsError() throws Exception {
         // Given...
         String propertyName = "property6";
         String value = "value6";
-        setServlet("/cps/framew0rk/properties/"+propertyName, "framework", value, "PUT");
+		JsonElement requestJson = JsonParser.parseString(String.format("{name: %s ,value: %s }",propertyName,value));
+		String requestBody = requestJson.toString();
+        setServlet("/cps/framew0rk/properties", "framework", requestBody, "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -290,7 +772,7 @@ public class TestPropertyRoute extends CpsServletTest{
 
         // When...
         servlet.init();
-        servlet.doPut(req, resp);
+        servlet.doPost(req, resp);
 
         // Then...
         assertThat(resp.getStatus()).isEqualTo(404);
@@ -305,201 +787,10 @@ public class TestPropertyRoute extends CpsServletTest{
     }
     
     @Test
-    public void TestPropertyRouteWithNamespaceNoValuePUTNewPropertyReturnsError() throws Exception {
+    public void TestPropertyRouteWithNamespaceNoRequestBodyPOSTNewPropertyReturnsError() throws Exception {
         // Given...
-        String propertyName = "property6";
-        String value = "";
-        setServlet("/cps/framew0rk/properties/"+propertyName, "framework", value, "PUT");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-        ServletOutputStream outStream = resp.getOutputStream();	
-
-        // When...
-        servlet.init();
-        servlet.doPut(req, resp);
-
-        // Then...
-        assertThat(resp.getStatus()).isEqualTo(411);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5411,
-            "E: Error occured when trying to access the endpoint '/cps/framew0rk/properties/property6'.",
-            " The request body is empty."
-		); 
-    }
-
-    @Test
-    public void TestPropertyRouteWithNamespaceNullValuePUTNewPropertyReturnsError() throws Exception {
-        // Given...
-        String propertyName = "property6";
-        String value = null;
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "PUT");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-        ServletOutputStream outStream = resp.getOutputStream();	
-
-        // When...
-        servlet.init();
-        servlet.doPut(req, resp);
-
-        // Then...
-        assertThat(resp.getStatus()).isEqualTo(411);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5411,
-            "E: Error occured when trying to access the endpoint '/cps/framework/properties/property6'.",
-            " The request body is empty."
-		); 
-    }
-
-    /*
-     * TESTS  -- POST requests
-     */
-    @Test
-    public void TestPropertyRouteGetPOSTFrameworkReturnsError() throws Exception{
-		// Given...
-		setServlet("/cps/namespace1/properties/property1",null ,"value12", "POST");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-				
-		// When...
-		servlet.init();
-		servlet.doPost(req,resp);
-
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5016,
-			"GAL5016E: Error occured when trying to access namespace 'namespace1'. The namespace provided is invalid."
-		);
-    }
-    
-    @Test
-    public void TestPropertyRoutePOSTBadNamespaceReturnsError() throws Exception{
-		// Given...
-		setServlet("/cps/error/properties/property1",null ,"value6", "PUT");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-				
-		// When...
-		servlet.init();
-		servlet.doPost(req,resp);
-
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5016,
-			"GAL5016E: Error occured when trying to access namespace 'error'. The namespace provided is invalid."
-		);
-    }
-
-    @Test
-    public void TestPropertyRouteWithExistingNamespacePOSTExistingPropertyReturnsOK() throws Exception {
-        // Given...
-        String propertyName = "property5";
-        String value = "value6";
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "POST");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-        ServletOutputStream outStream = resp.getOutputStream();	
-
-        // When...
-        servlet.init();
-        servlet.doPost(req, resp);
-
-        // Then...
-        Integer status = resp.getStatus();
-        String output = outStream.toString();
-        assertThat(status).isEqualTo(201);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-        assertThat(output).isEqualTo("Successfully updated property property5 in framework");
-        assertThat(checkNewPropertyInNamespace(propertyName, value)).isTrue();       
-    }
-
-    @Test
-    public void TestPropertyRouteWithExistingNamespacePOSTNewPropertyReturnsError() throws Exception {
-        // Given...
-        String propertyName = "property6";
-        String value = "value6";
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "POST");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-        ServletOutputStream outStream = resp.getOutputStream();	
-
-        // When...
-        servlet.init();
-        servlet.doPost(req, resp);
-
-        // Then...
-        Integer status = resp.getStatus();
-        assertThat(status).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-       checkErrorStructure(
-			outStream.toString(),
-			5017,
-			"E: Error occured when trying to access property 'property6'. The property name provided is invalid."
-		);        
-    }
-
-    @Test
-    public void TestPropertyRouteWithErroneousNamespacePOSTNewPropertyReturnsError() throws Exception {
-        // Given...
-        String propertyName = "property5";
-        String value = "value6";
-        setServlet("/cps/framew0rk/properties/"+propertyName, "framework", value, "POST");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-        ServletOutputStream outStream = resp.getOutputStream();	
-
-        // When...
-        servlet.init();
-        servlet.doPost(req, resp);
-
-        // Then...
-        assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-
-		checkErrorStructure(
-			outStream.toString(),
-			5016,
-			"E: Error occured when trying to access namespace 'framew0rk'. The namespace provided is invalid."
-		); 
-    }
-    
-    @Test
-    public void TestPropertyRouteWithNamespaceNoValuePOSTNewPropertyReturnsError() throws Exception {
-        // Given...
-        String propertyName = "property5";
-        String value = "";
-        setServlet("/cps/framew0rk/properties/"+propertyName, "framework", value, "POST");
+		String requestBody = "";
+        setServlet("/cps/framew0rk/properties", "framework", requestBody, "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -517,17 +808,16 @@ public class TestPropertyRoute extends CpsServletTest{
 		checkErrorStructure(
 			outStream.toString(),
 			5411,
-            "E: Error occured when trying to access the endpoint '/cps/framew0rk/properties/property5'.",
+            "E: Error occured when trying to access the endpoint '/cps/framew0rk/properties'.",
             " The request body is empty."
 		); 
     }
 
     @Test
-    public void TestPropertyRouteWithNamespaceNullValuePOSTNewPropertyReturnsError() throws Exception {
+    public void TestPropertyRouteWithNamespaceNullRequestBodyPOSTNewPropertyReturnsError() throws Exception {
         // Given...
-        String propertyName = "property6";
-        String value = null;
-        setServlet("/cps/framework/properties/"+propertyName, "framework", value, "POST");
+		String requestBody = null;
+        setServlet("/cps/framework/properties", "framework", requestBody, "POST");
 		MockCpsServlet servlet = getServlet();
 		HttpServletRequest req = getRequest();
 		HttpServletResponse resp = getResponse();
@@ -545,116 +835,8 @@ public class TestPropertyRoute extends CpsServletTest{
 		checkErrorStructure(
 			outStream.toString(),
 			5411,
-            "E: Error occured when trying to access the endpoint '/cps/framework/properties/property6'.",
+            "E: Error occured when trying to access the endpoint '/cps/framework/properties'.",
             " The request body is empty."
 		); 
-    }
-
-    /*
-     * TESTS  -- DELETE requests
-     */
-
-    @Test
-    public void TestPropertyRouteDELETENoFrameworkReturnsError() throws Exception{
-        // Given...
-		setServlet("/cps/namespace1/properties/property1", null, null, "DELETE");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-        
-		// When...
-		servlet.init();
-		servlet.doDelete(req,resp);
-        
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-        
-		checkErrorStructure(
-			outStream.toString(),
-			5016,
-			"GAL5016E: Error occured when trying to access namespace 'namespace1'. The namespace provided is invalid."
-        );
-    }
-        
-    
-    @Test
-    public void TestPropertyRouteDELETEBadNamespaceReturnsError() throws Exception{
-        // Given...
-		setServlet("/cps/error/properties/property1", null, null, "DELETE");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-        
-		// When...
-		servlet.init();
-		servlet.doDelete(req,resp);
-        
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-        
-		checkErrorStructure(
-            outStream.toString(),
-			5016,
-			"GAL5016E: Error occured when trying to access namespace 'error'. The namespace provided is invalid."
-            );
-        }
-        
-    @Test
-    public void TestPropertyRouteDELETEBadPropertyReturnsError() throws Exception{
-        // Given...
-		setServlet("/cps/framework/properties/badproperty", "framework", null, "DELETE");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-        
-		// When...
-		servlet.init();
-		servlet.doDelete(req,resp);
-        
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to Route
-		assertThat(resp.getStatus()).isEqualTo(404);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-        
-		checkErrorStructure(
-			outStream.toString(),
-			5017,
-			"E: Error occured when trying to access property 'badproperty'. The property name provided is invalid."
-        );
-    }
-
-    @Test
-    public void TestPropertyRouteDELETEPropertyReturnsOk() throws Exception{
-        // Given...
-        String propertyName = "property1";
-        String value = "value1";
-		setServlet("/cps/framework/properties/"+propertyName, "framework", null, "DELETE");
-		MockCpsServlet servlet = getServlet();
-		HttpServletRequest req = getRequest();
-		HttpServletResponse resp = getResponse();
-		ServletOutputStream outStream = resp.getOutputStream();	
-        
-		// When...
-		servlet.init();
-		servlet.doDelete(req,resp);
-        
-		// Then...
-		// We expect an error back, because the API server couldn't find any Etcd store to RouteInteger status = resp.getStatus();
-        String output = outStream.toString();
-        assertThat(resp.getStatus()).isEqualTo(200);
-		assertThat(resp.getContentType()).isEqualTo("application/json");
-		assertThat(resp.getHeader("Access-Control-Allow-Origin")).isEqualTo("*");
-        assertThat(output).isEqualTo("Successfully deleted property property1 in framework");
-        assertThat(checkNewPropertyInNamespace(propertyName, value)).isFalse();
     }
 }
